Fixes job denying.
Everytime you deny a job it checks for gangid in a faction job. Old dev probably copy pasted too happily and forgot where it's not needed. And likewise I just carry on the old way gangid are initiated. If someone figures this out in the future hope you trim it <3