Balance: Informatn quest and kill nato quest.
Informant went from 1500 to 4500 due to this is mostly a group quest and when initiated results in a lot of team work, for now it only pays the killer, but with a large sum maybe they'll be generous to their teammates.

Kill NATO quest is a very easy starting quest seeing how it's limiting in kill numbers, it's now doubled in per head reward.